introducing new library structure
