+++ data_ml_utils/config/config.json
@@ -1,0 +1,3 @@
+{
+    "s3_structure":["type", "project", "dataset", "use_case", "setup"]
+}
