+++ data_ml_utils/fselectutils.py
@@ -1,24 +1,32 @@
 """A library containing commonly used utils for machine learning feature selection
 """
-import logging
-import pandas as pd
-import csv
+from mainutils import main_utils
+from data_utils import generalutils as gu
 
-class general_metrics(object):
-    logger= None
-    metrics= {}
-    
-    def __init__(self):
-        self.logger= logging.getLogger(__name__)
-        self.logger.setLevel(logging.CRITICAL)
+from .config import load_config
+config = load_config()
+
+class general_select(main_utils):
+    logger_name='general_select'
+    feature_metrics= {}
 
     def get_features_corr_matrix(self, matrix):
-        x_pd= pd.DataFrame(matrix)
+        x_pd= gu.create_data_frame(matrix)
         corr_mtx= x_pd.corr()
         corr_mtx.index.name ='feature'
         corr_mtx.reset_index(level=0, inplace=True)
-        self.metrics['corr_matrix'] = corr_mtx
+        self.feature_metrics['corr_matrix'] = corr_mtx
+    
+    def export_metrics(self):
+        if self.export_local:
+            self.export_corr_matrix_as_text()
+        if self.export_s3:
+            self.export_corr_matrix_to_s3()
     
     def export_corr_matrix_as_text(self):
-        df= self.metrics['corr_matrix']
-        df.to_csv(r'feature_corr_matrix.csv', index = False, header=True)
+        df= self.feature_metrics['corr_matrix']
+        self.export_df_as_text(df, 'feature_corr_matrix')
+    
+    def export_corr_matrix_to_s3(self):
+        df= self.feature_metrics['corr_matrix']
+        self.export_metric_to_s3(df, 'feature_corr_matrix', 'feature_corr_matrix')
